Use correct Python for running scss
After upgrading django_compressor (even if just upgrading to 1.6), tests started
failing with this while running scss:

/usr/bin/python: No module named scss

This is because the Docker environment is using /usr/local/bin/python, so likely
this step never actually worked. Looking at the compressor changelog:

https://github.com/django-compressor/django-compressor/compare/1.5...1.6

This commit seems relevant:

https://github.com/django-compressor/django-compressor/commit/85a190090f781a8617401050a91d5530500b94e2

Before that comment, this error probably was hidden? Not sure why/how scss
compilation worked then, though...

With this change, the same Python Django is executed with is now used to call
scss as well.