Python 3: Fix rounding of lecturer ratings
On Python 2, round() with a x.5 number always rounds up.

On Python 3, "banker's rounding" / "round to even" is done, which always rounds
to even numbers in order to not accumulate rounding errors.

However, for ratings that's not what we want, instead we always need to round
up. Doing int(x + 0.5) is the easiest way to do so. It only works for positive
numbers, but that's fine for ratings.

For more than you'd ever wanted to know about rounding in Python, see:

https://realpython.com/python-rounding/
https://stackoverflow.com/questions/33019698/