Add tests for content-disposition header
Note the old testUmlautDocumentServed test didn't actually work. It tried to set
original_filename on the document, but that attribute is overwritten via the
upload_to argument of "document". It calls the document_file_name function,
which then sets instance.original_filename.

The filename model-bakery uses is hardcoded to be a timestamp, so we instead
create the object by hand here.

See #217, #241